# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEC",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/",
            "description": "DEV"
        }
    ],
    "paths": {
...
        "/groups/{groupId}/members": {
            "get": {
                "tags": [
                    "members"
                ],
                "summary": "Listar todos os membros",
                "description": "Lista todos os membros: ADMINISTRADOR, REPRESENTANTE E GERENTE têm acesso a este endpoint.",
                "operationId": "de831537748e1740e1890ce311914e44",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "O ID do grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "members"
                ],
                "summary": "Criar novo membro",
                "description": "Cria um novo membro, somente o REPRESENTANTE tem acesso a este endpoint.",
                "operationId": "f2355c9ee72792388c85bcb3b7f64797",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id do grupo que o membro sera associado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": [
                                    {
                                        "email": "bar@mail.com",
                                        "role": "bar",
                                        "phone": "93991185489",
                                        "entry_date": "23-10-1998",
                                        "departure_date": "23-10-2023"
                                    },
                                    {
                                        "email": "outromail@mail.com",
                                        "role": "bar",
                                        "phone": "93991185489",
                                        "entry_date": "23-10-1998",
                                        "departure_date": "23-10-2023"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/groups/{groupId}/members/{id}": {
            "get": {
                "tags": [
                    "members"
                ],
                "summary": "Lista o registro de membro por ID",
                "description": "Lista o registro de membro por ID de referência.",
                "operationId": "8817ebb1aaf5600c07f5d1d347f76bf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do membro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Membro not found"
                    },
                    "200": {
                        "description": "Ok"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            },
            "put": {
                "tags": [
                    "members"
                ],
                "summary": "Atualizar membro",
                "description": "somente o REPRESENTANTE tem acesso a este endpoint.",
                "operationId": "af90c926d353b3ee7065a0330587d36b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do membro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "O ID do grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "description": "O papel do membro.",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "O número de telefone do membro (deve ter pelo menos 11 caracteres).",
                                        "type": "string"
                                    },
                                    "departure_date": {
                                        "description": "A data de partida do membro (formato YYYY-MM-DD).",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "description": "Email do membro.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Membro not found"
                    }
                }
            }
        },
        "/groups/{groupId}/members/{memberID}": {
            "delete": {
                "tags": [
                    "members"
                ],
                "summary": "Deletar membro",
                "description": "Deletar membro por ID de referência, somente o REPRESENTANTE tem acesso a este endpoint.",
                "operationId": "4f9be8ba174845db12a096823062ecfa",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Id do grupo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do membro",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Erro"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}